Improve application question check
The check was failing if only require_application was passed
and not application_question, despite application_question
being set in database.
